solve array of strings to array of numbers

+++ templatetest/arrayofstringstoarrayofnumbers.cpp
@@ -18,7 +18,7 @@ Note that you can receive floats as well.
 	std::vector<float> result;
 	for(auto c:arr)
 		{
-			result.push_back(stoi(c));
+			result.push_back(stof(c));
 			std::cout << c << std::endl;
 		}
 	return result;
